Add a mechanism for skipping thumbnail generation
Its a workaround for a problem where the job won't time out so it
locks up the job queue.

+++ rails/app/jobs/generate_thumbnail_job.rb
@@ -5,7 +5,19 @@ class GenerateThumbnailJob < ApplicationJob
   def perform(model_name, site_id)
     return unless model_name.in?(%w[ Site TspotSite ])
     return unless site = model_name.safe_constantize&.find_by_id(site_id)
+
+    # Fixme: Not sure why but there are some sites that choke the job
+    # queue and somehow don't time out
+    return if site_id.in?(Settings.secrets(:skip_thumbnail_sites, model_name.to_sym)||[])
+
+    # If the thumbnail is new enough then no need to regenerate
+    return if site.thumbnail_fresh?
+
+    # Remove queued jobs that would regenerate the thumbnail we're about to create
+    # (This might be ineffective if the worker loaded them already.)
     dupe_jobs(model_name, site_id).delete_all
+
+    # Generate thumbnail
     site.send(:update_thumbnail_now)
   end
